requirements.txt
yfinance==0.2.36
kafka-python==2.0.2
pyspark==3.5.0
delta-spark==3.0.0
apache-airflow==2.8.1
apache-airflow-providers-amazon==8.14.0
boto3==1.34.0
pytest==7.4.4
pytest-cov==4.1.0
